вынесла подключение к db в отдельный файл
